refactor: 도메인 특화 컴포넌트 designsystem -> ui 모듈로 이동
