+++ core/ui/src/main/kotlin/com/useai/core/ui/LogitCtaButton.kt
@@ -1,4 +1,4 @@
-package com.useai.core.designsystem.component.button
+package com.useai.core.ui
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -6,6 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.useai.core.designsystem.component.button.LogitPrimaryButton
 import com.useai.core.designsystem.theme.LogitTheme
 
 @Composable
