+++ core/ui/src/main/kotlin/com/useai/core/ui/LogitFormTitle.kt
@@ -1,4 +1,4 @@
-package com.useai.core.designsystem.component
+package com.useai.core.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
