+++ feature/newproject/src/main/kotlin/com/useai/feature/newproject/ui/NewProjectBasicInfo.kt
@@ -22,11 +22,11 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.slack.circuit.codegen.annotations.CircuitInject
 import com.useai.core.designsystem.R
-import com.useai.core.designsystem.component.LogitFormTitle
-import com.useai.core.designsystem.component.LogitInputField
+import com.useai.core.ui.LogitFormTitle
+import com.useai.core.ui.LogitInputField
 import com.useai.core.designsystem.component.appbar.PopUpTitle
-import com.useai.core.designsystem.component.button.LogitCtaButton
-import com.useai.core.designsystem.component.stepper.LogitStepper
+import com.useai.core.ui.LogitCtaButton
+import com.useai.core.ui.LogitStepper
 import com.useai.core.designsystem.theme.LogitTheme
 import com.useai.feature.newproject.NewProjectBasicInfoScreen
 import dagger.hilt.android.components.ActivityRetainedComponent
